+++ App.js
@@ -3,7 +3,7 @@ import { createStackNavigator } from "@react-navigation/stack";
 import * as React from "react";
 import { Platform, StatusBar, StyleSheet, View } from "react-native";
 
-import useCachedResources from "./hooks/useCachedResources";
+import { useCachedResources } from "./+hooks/useCachedResources";
 import BottomTabNavigator from "./navigation/BottomTabNavigator";
 import LinkingConfiguration from "./navigation/LinkingConfiguration";
